Add FaviconController
